:star: Functional package refactor
