+++ packages/mechanic-canvas-engine/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@designsystemsinternational/mechanic-canvas-engine",
+  "version": "0.0.1",
+  "main": "index.js",
+  "scripts": {
+  },
+  "dependencies": {
+    "mechanic-utils": "../mechanic-utils"
+  }
+}
