+++ packages/mechanic-p5-engine/index.js
@@ -1,0 +1,32 @@
+import { Mechanic } from "mechanic-utils";
+import p5 from "p5";
+
+const root = document.getElementById("root");
+
+let p5Sketch;
+
+const run = (functionName, func, values, isPreview) => {
+  if (p5Sketch) {
+    p5Sketch.remove();
+  }
+  const mechanic = new Mechanic(func.params, func.settings, values);
+  const onFrame = () => {
+    if (!isPreview) {
+      mechanic.frame(root.childNodes[0]);
+    }
+  };
+  const onDone = async () => {
+    p5Sketch.noLoop();
+    if (!isPreview) {
+      await mechanic.done(root.childNodes[0]);
+      mechanic.download(functionName);
+    }
+  };
+  p5Sketch = new p5(
+    (sketch) =>
+      func.handler(sketch, mechanic.values, { frame: onFrame, done: onDone }),
+    root
+  );
+};
+
+export default run;
