+++ packages/mechanic-template/app/App.js
@@ -1,12 +1,13 @@
 import React from "react";
 import { Switch, Route, withRouter } from "react-router-dom";
-import { requireFunctions } from "./utils";
+import { requireFunctions } from "mechanic-utils";
 
 import Function from "./pages/Function";
 import NotFound from "./pages/NotFound";
 import Nav from "./components/Nav";
 
-const functions = requireFunctions();
+const functionContext = require.context("../functions", true, /^(.{2,})\/index\.js$/);
+const { functions } = requireFunctions(functionContext);
 const functionNames = Object.keys(functions);
 
 import css from "./App.css";
@@ -22,7 +23,7 @@ const App = props => {
             path={i == 0 ? ["/", `/${name}`] : `/${name}`}
             render={() => (
               <Function name={name} exports={functions[name]}>
-                <Nav name={name} functions={functions} />
+                <Nav name={name} functionsNames={functionNames} />
               </Function>
             )}
           />
