+++ packages/mechanic-template/functions/randomImage/index.js
@@ -1,3 +1,5 @@
+import engine from "mechanic-canvas-engine";
+
 export const handler = (params, mechanic) => {
   const { width, height, background, color1, color2, numberOfSquares } = params;
   const colors = [color1, color2];
@@ -47,6 +49,6 @@ export const params = {
 };
 
 export const settings = {
-  engine: "canvas",
+  engine,
   usesRandom: true
 };
