+++ packages/mechanic-template/functions/reactVideo/index.js
@@ -1,4 +1,5 @@
 import React, { useState, useEffect, useRef } from "react";
+// import engine from "mechanic-react-engine";
 
 export const handler = ({ width, height, frame, done, background, fill }) => {
   const isPlaying = useRef(true);
@@ -47,7 +48,7 @@ export const params = {
 };
 
 export const settings = {
-  engine: "react",
+  // engine,
   animated: true
 };
