+++ packages/mechanic-template/functions/svgImage/index.js
@@ -33,7 +33,4 @@ export const params = {
   }
 };
 
-export const settings = {
-  returns: "svgString",
-  type: "image"
-};
+export const settings = {};
