+++ packages/mechanic-template/functions/svgVideo/index.js
@@ -49,6 +49,5 @@ export const params = {
 };
 
 export const settings = {
-  returns: "svgString",
-  type: "video"
+  animated: true
 };
