+++ packages/mechanic-template/package.json
@@ -46,6 +46,10 @@
   "dependencies": {
     "mousetrap": "^1.6.5",
     "p5": "^1.1.9",
-    "seedrandom": "^3.0.5"
+    "seedrandom": "^3.0.5",
+    "mechanic-utils": "../mechanic-utils",
+    "mechanic-canvas-engine": "../mechanic-canvas-engine",
+    "mechanic-p5-engine": "../mechanic-p5-engine",
+    "mechanic-react-engine": "../mechanic-react-engine"
   }
 }
