+++ packages/mechanic-utils/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@designsystemsinternational/mechanic-utils",
+  "version": "0.0.1",
+  "main": "src/index.js",
+  "scripts": {
+  },
+  "dependencies": {
+    "seedrandom": "^3.0.5"
+  }
+}
