+++ packages/mechanic-utils/src/index.js
@@ -1,0 +1,2 @@
+export { requireFunctions, setUp } from "./function-set-up";
+export { Mechanic } from "./mechanic";
