+++ packages/mechanic-utils/src/mechanic-error.js
@@ -1,0 +1,8 @@
+class MechanicError extends Error {
+  constructor(message) {
+    super(message);
+    this.name = "MechanicError";
+  }
+}
+
+export { MechanicError };
